Face_regcognition/requirements.txt
alembic==1.6.5
aniso8601==8.0.0
attrs==20.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.1
cycler==0.10.0
dataclasses==0.8
dataset==1.1.2
dlib==19.22.0
easydict==1.9
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==0.1.0
Flask==1.1.1
flask-restplus==0.13.0
greenlet==1.1.0
idna==3.2
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.13.2
jsonschema==3.2.0
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.0.0
numpy==1.17.0
opencv-python==4.2.0.34
pandas==1.1.5
Pillow==6.1.0
protobuf==3.17.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2019.3
requests==2.26.0
scikit-learn==0.21.3
scipy==1.5.4
seaborn==0.9.0
six==1.14.0
SQLAlchemy==1.4.22
tensorboardX==2.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.31.1
typing-extensions==3.7.4.1
urllib3==1.26.6
uWSGI==2.0.18
Werkzeug==0.15.2
zipp==3.4.1
